move PackageUtils to ProjectAuditor/Editor
add InstallPackage, DownloadFile and Progress  static methods
